+++ app/src/main/java/com/haque/baseline/data/source/source/repository/WeatherRepositoryImpl.kt
@@ -6,7 +6,7 @@ import com.haque.baseline.domain.WeatherRepository
 import javax.inject.Inject
 
 class WeatherRepositoryImpl @Inject constructor (private val api: WeatherApi): WeatherRepository {
-    override suspend fun getOneCallAPIResponse(lat: Double, lng: Double): OneCallWeatherPayloadDTO {
+    override suspend fun getOneCallAPIResponse(lat: Float, lng: Float): OneCallWeatherPayloadDTO {
         return api.getWeatherData(lat, lng)
         // DO NOT keep try catches in network calls.  Try Catch's are slow
         // and you should not slow down processes that require frequent or semi - frequent network calls
